[feat] agrego el buscador de ordenes de compras en el timeline de los leads
