+++ leads/app/controllers/home.php
@@ -1,24 +1,29 @@
 <?php 
-
+use Enid\BusquedaRecibo\Form as FormBusquedaOrdenesCompra;
 class Home extends CI_Controller
 {
-    
+    private $form_busqueda_ordenes_compra;
     function __construct()
     {
         parent::__construct();                       
         $this->load->helper("busqueda");
-        $this->load->library(lib_def());        
+        $this->load->library(lib_def());  
+        $this->form_busqueda_ordenes_compra = new FormBusquedaOrdenesCompra();      
         
     }
 
     function index()
     {
+        $param = $this->input->get();
         $data = $this->app->session();
-        $data = $this->app->cssJs($data, "leads");                            
+        $data = $this->app->cssJs($data, "leads");                                    
+        $data["leads"] = $this->recibos_sin_ficha_seguimiento();    
+        $data["status_ventas"] = $this->app->api("status_enid_service/index");    
+        $data["tipos_entregas"]= $this->app->api("tipo_entrega/index/");
         
-        $data["leads"] = $this->recibos_sin_ficha_seguimiento();        
+        $data["formulario_busqueda_ordenes_compra"] = 
+        $this->form_busqueda_ordenes_compra->busqueda($param, $data,1);
         $this->app->pagina($data, render($data), 1);
-
 
     }
     function recibos_sin_ficha_seguimiento()
