+++ pedidos/app/helpers/pedidos_helper.php
@@ -1711,135 +1711,11 @@ if (!function_exists('invierte_date_time')) {
 
         return $response;
     }
-    function fechas_busqueda()
-    {
-        $fechas[] =
-            [
-                "fecha" => "FECHA REGISTRO",
-                "val" => 1,
-            ];
-        $fechas[] =
-            [
-                "fecha" => "FECHA CONTRA ENTREGA",
-                "val" => 5,
-            ];
-        $fechas[] =
-            [
-                "fecha" => "FECHA ENTREGA",
-                "val" => 2,
-            ];
-        $fechas[] =
-            [
-                "fecha" => "FECHA CANCELACION",
-                "val" => 3,
-            ];
-        $fechas[] =
-            [
-                "fecha" => "FECHA PAGO",
-                "val" => 4,
-            ];
-        return $fechas;
-    }
-    function form_busqueda_ordenes_compra($param, $data)
+    
+    function get_form_busqueda_pedidos($data)
     {
 
-        $es_busqueda_reparto = prm_def($param, 'reparto');
-        $ancho_fechas = 'col-sm-6 mt-5 p-0 p-md-1 ';
-        $tipos_entregas = $data["tipos_entregas"];
-        $status_ventas = $data["status_ventas"];
-
-        $response[] = form_open("", ["class" => "form_busqueda_pedidos mt-5", "method" => "post"]);
-        $response[] = hiddens(['class' => 'usuarios', 'name' => 'usuarios', 'value' => prm_def($param, 'usuarios')]);
-        $response[] = hiddens(['class' => 'ids', 'name' => 'ids', 'value' => prm_def($param, 'ids')]);
-        $response[] = hiddens(['class' => 'es_busqueda_reparto', 'name' => 'es_busqueda_reparto', 'value' => $es_busqueda_reparto]);
-        $response[] = hiddens([
-            'class' => 'es_administrador', 'name' => 'es_administrador',
-            'value' => es_administrador($data)
-        ]);
-
-        $response[] = hiddens(
-            [
-                "name" => "perfil",
-                "class" => "perfil_consulta",
-                "value" => $data["id_perfil"],
-            ]
-        );
-
-        $select_comisionistas = create_select(
-            $data['comisionistas'],
-            'id_usuario_referencia',
-            'comisionista form-control',
-            'comisionista',
-            'nombre_usuario',
-            'id',
-            0,
-            1,
-            0,
-            '-'
-        );
-
-        $response[] = form_busqueda_pedidos($data, $tipos_entregas, $status_ventas, fechas_busqueda());
-        $es_busqueda = keys_en_arreglo(
-            $param,
-            [
-                'fecha_inicio',
-                'fecha_termino',
-                'type',
-                'servicio'
-            ]
-        );
-
-        $visibilidad = (!es_data($data['comisionistas'])) ? 'd-none' : '';
-        $response[] = flex_md(
-            'Filtrar por vendedor',
-            $select_comisionistas,
-            _text_('col-sm-12 mt-md-m5 mt-3 p-0', _between_md, $visibilidad),
-            _text_('text-left', $visibilidad),
-            _text_('text-left', $visibilidad)
-        );
-
-        if ($es_busqueda && es_data($param)) {
-
-            $response[] = frm_fecha_busqueda(
-                $param["fecha_inicio"],
-                $param["fecha_termino"],
-                $ancho_fechas,
-                $ancho_fechas
-            );
-            $response[] = hiddens(
-                [
-                    "name" => "consulta",
-                    "class" => "consulta",
-                    "value" => 1,
-                ]
-            );
-            $response[] = hiddens(
-                [
-                    "name" => "servicio",
-                    "class" => "servicio",
-                    "value" => $param["servicio"],
-                ]
-            );
-            $response[] = hiddens(
-                [
-                    "name" => "type",
-                    "class" => "type",
-                    "value" => $param["type"],
-                ]
-            );
-        } else {
-
-            $response[] = frm_fecha_busqueda(0, 0, $ancho_fechas, $ancho_fechas);
-        }
-
-        $response[] = form_close();
-        return d($response, 12);
-    }
-    function get_form_busqueda_pedidos($data, $param)
-    {
-
-        $z[] = d(form_busqueda_ordenes_compra($param, $data), 13);
-        $z[] = place("place_pedidos ");
+        $z[] = d($data["formulario_busqueda_ordenes_compra"], 13);        
         $z[] = frm_busqueda();
 
         $response[] = d($z);
@@ -2689,96 +2565,7 @@ if (!function_exists('invierte_date_time')) {
         return append($r);
     }
 
-    function visibilidad_input_form($data)
-    {
-        return es_cliente($data) ? 'd-none' : '';
-    }
-    function form_busqueda_pedidos($data, $tipos_entregas, $status_ventas, $fechas)
-    {
-
-        $id_perfil = $data['id_perfil'];
-        $restriciones_comisionista_busqueda = $data['restricciones']['restriccion_status_comisionista_busqueda'];
-        $restricciones_administrador_busqueda = $data['restricciones']['restricciones_administrador_busqueda'];
-        $restricciones = ($id_perfil == 6) ? $restriciones_comisionista_busqueda : [];
-        $restricciones = (!in_array($id_perfil, [20, 6])) ? $restricciones_administrador_busqueda : $restricciones;
-        $visibilidad = visibilidad_input_form($data);
-
-        $r[] = input_frm(
-            _text_(_6p, 'mt-5', $visibilidad),
-            'Cliente',
-            [
-                "name" => "cliente",
-                "id" => "cliente",
-                "placeholder" => "Nombre, correo, telefono ...",
-                "class" => "input_busqueda",
-                "onpaste" => "paste_busqueda();"
-            ]
-        );
-
-
-        $r[] = hiddens(
-            [
-                "name" => "v",
-                'value' => 1,
-            ]
-        );
-
-        $r[] = input_frm(
-            _text_(_6p, 'mt-5'),
-            '#Recibo',
-            [
-                "name" => "recibo",
-                "id" => 'busqueda_recibo'
-            ]
-        );
-
-
-        $r[] = d(flex(
-            "Tipo de entrega",
-            create_select(
-                $tipos_entregas,
-                "tipo_entrega",
-                "tipo_entrega form-control",
-                "tipo_entrega",
-                "nombre",
-                "id",
-                0,
-                1,
-                0,
-                "-"
-            ),
-            "flex-column col-md-4 p-0 mt-3"
-        ), $visibilidad);
-
-        $r[] = d(flex(
-            'Status',
-            create_select(
-                $status_ventas,
-                "status_venta",
-                "status_venta  form-control",
-                "status_venta",
-                "text_vendedor",
-                "id_estatus_enid_service",
-                0,
-                1,
-                0,
-                "-",
-                $restricciones
-            ),
-            "flex-column col-md-4 p-0 mt-3"
-        ), $visibilidad);
-
-
-        $busqueda_orden = create_select_selected($fechas, 'val', 'fecha', 5, 'tipo_orden', 'tipo_orden form-control');
-        $r[] = d(flex(
-            "Ordenar",
-            $busqueda_orden,
-            "flex-column col-md-4 p-0 mt-3"
-        ), $visibilidad);
-
-
-        return append($r);
-    }
+    
 
 
     function puntos_encuentro(
